requirements.txt
Flask>=2.2.5
Flask-WTF==1.0.0
requests==2.26.0
gunicorn==20.1.0
Flask-SQLAlchemy
Flask-Migrate
python-dotenv
